--- src/apply.py
@@ -1,2 +1,0 @@
-def apply_points(row):
-    return row.win*3+row.row
